+++ package.json
@@ -14,6 +14,7 @@
     "@expo/vector-icons": "^13.0.0",
     "@mui/icons-material": "^5.15.15",
     "@mui/material": "^5.15.15",
+    "@react-native-community/blur": "^4.4.0",
     "@react-navigation/bottom-tabs": "^6.5.11",
     "@react-navigation/elements": "^1.3.21",
     "@react-navigation/native": "^6.1.9",
